+++ lib/src/future_handler.dart
@@ -1,8 +1,6 @@
-import 'dart:io';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
+import 'package:jin_widget_helper/src/jin_widget.dart';
 import '../jin_widget_helper.dart';
 
 class FutureHandler<T> extends StatelessWidget {
@@ -36,7 +34,7 @@ class FutureHandler<T> extends StatelessWidget {
             return Center(child: error(snapshot.error.toString()));
           return Center(child: Text(snapshot.error.toString()));
         } else {
-          return Center(child: loading ?? UIHelper.loadingChild);
+          return Center(child: loading ?? JinWidget.loadingChild);
         }
       },
     );